feat(profile): allow fractional years of experience (e.g. 2.5)
- years_experience: Integer -> Numeric(4,1) (asdecimal=False, plain float)
- schema int -> float; migration c2d3e4f5a6b7 (widen, lossless)
- Years-of-experience input: step=0.5, inputMode=decimal, placeholder
- Seed hasnainrazaa03 brag bank with 27 grounded facts
- TESTING: T-BRAG-05/06 pass; log issue #7 (years decimal)

+++ backend/app/schemas/profile.py
@@ -28,7 +28,7 @@ class CareerProfileUpdate(ProfileModel):
 
     work_authorization: str | None = Field(default=None, max_length=120)
     visa_status: str | None = Field(default=None, max_length=120)
-    years_experience: int | None = Field(default=None, ge=0, le=60)
+    years_experience: float | None = Field(default=None, ge=0, le=60)
 
     preferred_roles: list[str] | None = Field(default=None, max_length=20)
     preferred_industries: list[str] | None = Field(default=None, max_length=20)
@@ -63,7 +63,7 @@ class CareerProfileResponse(ProfileModel):
     website_url: str | None
     work_authorization: str | None
     visa_status: str | None
-    years_experience: int | None
+    years_experience: float | None
     preferred_roles: list[str]
     preferred_industries: list[str]
     preferred_locations: list[str]
